+++ helpers/handleMongooseError.js
@@ -5,13 +5,3 @@ const handleMongooseError = (error, data, next) => {
 };
 
 module.exports = handleMongooseError;
-
-// const errorHandler = (error, req, res, next) => {
-//   const {
-//     message = "Something went wrong please try again later",
-//     statusCode = 500,
-//   } = error;
-//   res.status(statusCode).json({ message });
-// };
-
-// module.exports = errorHandler;
