+++ routes/api/contacts.js
@@ -4,7 +4,7 @@ const ctrl = require("../../controllers/contacts");
 
 const { validateBody, isValidId } = require("../../middlewares");
 
-const { schemas } = require("../../models/contact");
+const { schemas } = require("../../utils/validation/contactValidationSchemas");
 
 const router = express.Router();
